[CMPLRLLVM-29178] Avoid HIRLMM-specific assumptions in areDDEdgesLegal
HIRLMM's invariance check has an exception for memrefs equal to the one
being queried, because in the context of HIRLMM those would be
hoisted/sunk along with the queried memref. However, that's not the case
for other uses of the invariance check that are possible now that it's
been exposed via HIRTransformUtils::isLoopInvariant, and this has caused
some serious issues in HIRRowWiseMV. This change replaces that
equal-memref exception with a specific check against the memrefs
recorded in the hoist/sink group so that the check is still valid even
if the hoist/sink group doesn't include all memrefs equal to the queried
one.

Change-Id: I3962be69d203829ea0611cddbfb919921fdb535d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/331654
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Wenju He <wenju.he@intel.com>