[NFC] Adding new SECRET feature SW_ADVANCED
We have a project that requires source sharing and we want to strip out
specific optimizations from this sharing. These features are expected to
be shipped in product.

Change introduces new macro and adds it enabled in all compiler builds
by default. There will be special build which will have this macro disabled.

Change-Id: I4831767595fad5bae38fc060db12d4a39abbc158
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/331679
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>