[2/2][VPlan][SVA] Sync SVA with CG for handling masked stores to uniform addresses.
Currently, VPlan vector CG optimizes stores to uniform address
only when they are unmasked. This patch fixes a bug in SVA to
synchronize its results with CG's behavior for such stores.
Before this fix SVA was marking these stores as LastPrivate
which causes stability issues when the SVA results are used in
CG.

Change-Id: I80623823af967be25bdafe11c49f62b55395c2ea
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/331868
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>