[Driver] Add support for Windows debug libraries with -fsycl -Qmkl -MDd
New debug libraries have been introduced for Windows builds of MKL.
These libraries (mkl_sycld and mkl_tbb_threadd) should be pulled
in with -Qmkl -MDd -fsycl.

Do a refactor of how the performance libraries are pulled in.  Instead
of adding the libs to the command line early in the compile, add at
the end during the linking phase, which better matches expectations and
does not fill the command line with non-user based inputs.

Contains additional fix for paren failure check on Windows debug

Change-Id: Ic1992c205919bf9b98bb3e5dc1328557a1f1220a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/331884
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Chilamantula, Harini <harini.chilamantula@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>