Revert "Remove deprecation markers for CreateLoad and CreateAlignedLoad"
This reverts commit 7630f8e71b20cad762d87fc2e9ce61e9231b2649.
The original commit to remove depreciaton markers was required to
workaround a build failure caused due to depreciation of specific
signatures of CreateLoad and CreateAlignedLoad. This depreciation was
done in the LLVM community as part of work towards support of opaque
pointers. 
This change reintroduces the depreciation markers.
Please look at CMPLRLLVM-27718 for further details.
There was one build failure caused due to this revert and has been
fixed here.
Change-Id: Idedd53d313ab614cec77f0a0020324ecc4034632
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/332092
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>