CMPLRLLVM-29320: Fix to not use getNextNode() in parser to find pragma loop.
The implementation of getNextNode() depends on availability of top sort
numbers which are not assigned until the end of framework.

There is another instance of getNextNode() in parser which needs to be
removed. The utilities should assert if the caller is trying to use
top sort number of 0 which is the default initialization value.

Change-Id: I4b06c2d8a11dbeb1ae93540f22ccf8d7b18fc098
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/332151
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Choi, Yoonseo <yoonseo.choi@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>