[OpenCL 3.0] Fix incorrect built-in lowering in offline mode
Some built-ins are not lowered correctly when linking SPIR-V program with the
option "-cl-std=CL3.0". It is because only "CL2.0" is handled.

For builin-in lowering, OpenCL C 3.0 is back compatible with OpenCL C 2.0 in our
implementation, we can reuse the lowering mechanism of OpenCL C 2.0.

BTW, add "CL3.0" pattern for CompilationUtils tool.

Change-Id: I1b1301d5a0a633c0a78fc7af92d6ddd9a3ab8450
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/332215
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>
Reviewed-by: Wenju He <wenju.he@intel.com>