[OpenCL] Fix calling convention issue
The calling convention is not set for newly created call instruction in
subgroup emulation pass and ResolveSubgroupWICall pass. If it's not same
between the function and the call instruction, instcombine and
SimplifyCFG pass will emit unreachable instruction which will trigger illegal
instruction error on x86.
Fixed by setting the calling convention of call instruction same as the older
one.

Change-Id: I64f4ea4b38f11692932d8d1f585d04247d7f3c4c
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/332282
Reviewed-by: Zou, Feng <feng.zou@intel.com>
Tested-by: Zou, Feng <feng.zou@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>