[VPlan] Fix privates source lookup and handling of lifetime/start/end routines with PHI as input argument.
This patch fixes the routine which traverses up the use-def chain to
check if a particular pointer is an alias of a private-memory. We
skipped check for PHI-instruction. This has been fixed and the existing
recursive lookup has been replaced by a Worklist-based algorithm.

We also add an additional check in vectorizeVPPHINode where we exclude
PHIs which mix SOA-pointers and are exlusively used in
lifetime.start/end intrinsics.

JIRA: CMPLRLLVM-29321

Change-Id: Id98bc37d2d90cc5d52bd85582c0d3d819d8ce5f2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/332422
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>