CMPLRLLVM-28297: Simplified IPO dead arguments elimination
This change set introduces a simplified dead argument elimination.
It will analyze when an argument is used only for store information
and it won't be used anymore. The actual parameter can be removed
if we can prove through IPO that it won't be used anymore. This is
a more simplified dead argument elimination compared to the
traditional dead argument elimination, which is very aggressive.
The intention is to add this pass late in the LTO process in
order to clean the IR before loop-opt.

Change-Id: Ifd51804259683aae2424f0fe7c7b8200119bc5a6
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/334314
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>