CMPLRLLVM-28847: Include sincos when generating math declarations for libimf
This change set adds the identification of possible sincos operation.
If the module has the functions sin and cos, there is a chance
that there is a computation for sin and cos of the same angle. In
that case, the compiler may convert the calls to both functions
into a call to sincos (optimized function). We would like to
use sincos from libimf rather than libm. This change set also
includes fmod. Performance restored:

    sycl_benchmarks_omp/aobench     +55.34%

Possible new gains (not confirmed if it is a regression)
    Linux: cpu2017/521.wrf_r        +2.03%
    Windows: cou2017/538.imagick_r  +5.43%

Change-Id: Ib39b7821995379486ba929ea926f0a3cbd8d8993
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/333044
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>