[VPlan][Minmax+index]Account [possibly] negative coefficient on linears.
The current processing of linear indexes does not account
negativeness of the indexes. This patch disables recognition
of such indexes as linear ones which will enforce emitting an
additional linear index (if it does not exist). The check for
both negative constant and blob indexes is performed. Blob indexes
are unknown to positive or negative so we should be conservative.

Fix for cmplrllvm-29216.

Change-Id: Icd3dc181fd2af63f454a300e54b154b04d22de4a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/332958
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>