CMPLRLLVM-29570: Fix seg fault when classic inlining report is run on 527.cam4
The problem was that instcombine removed some scoping intrinsics. When the
inlining report callback was executed, it did not null out the CallBase
instruction in the corresponding InlineReportCallSite. Then, cloning after
inlining occurred, accessing the now deleted CallBase caused a seg fault.

The LIT test only tests the classic inlining report because support for
cloning after inlining is not yet implemented for the metadata inlining report.

Change-Id: I3169a67d3df557c20f818f9c28e165bd10b1c12a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/333090
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>