Implement DTrans Metadata info.
See documentation/design here:
llvm/llvm/docs/Intel/DTrans/dtrans-opaque-pointers.rst

DTrans metadata is put in place to make sure that the DTrans
optimization pass has the correct amount of information to do its
optimizations in light of the OpaquePtr changes.  This patch is an
experimental implementation of that metadata generation that intends,
and is only enabled specifically for SPEC benchmarks.

Change-Id: I9d910d144780b0053a5e5288ab9f09ceb6eec0a8
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/334196
Reviewed-by: Keane, Erich <erich.keane@intel.com>
Tested-by: Keane, Erich <erich.keane@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>