[DTrans - OpaquePtr] Implement AOS-to-SOA type creation
This implements the prepareTypes() and populateTypes() methods to create the
new structure for the structures being transformed. This also requires some
enhancements to the DTransOPOptBase class to recognize that a type dependency
will exist when a pointer type is converted to a non-pointer type for structures
types that had a pointer to the original type as a field member.

Change-Id: I151fa7d8792a94a0b18732d153ccb08a28e7ad87
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/333702
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>