[DTrans - OpaquePtr] AOS-to-SOA pass transformation for SelectInst/PHINode
Handle cases of select/phi instructions for the AOS-to-SOA pass that
supports opaque pointers. When the result type for these instructions
are effectively pointers to the type being transformed, these cases
need to be recognized to convert them integer index values, and any
'null' values need to be recognized as the index 0.

Change-Id: I120b1abe89d5b5eeef27a24e1178c531308079b2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/333708
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>