[DTrans - OpaquePtr] AOS-to-SOA pass transformation for memory allocation and free
Implements the change to the allocation and free call sites for the array
of structures being transformed. This rewrites the IR to replace the original
allocation and free, to instead allocate a structure of arrays, and initialize
the pointer addresses within this structure to be subdivisions of the allocated
memory block.

Change-Id: I2f7b36034a51a34bc9d892914331e65223bd8e64
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/333710
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Mikhail Lychkov <mikhail.lychkov@intel.com>