[NFC] Add an option to add LSR to pipeline for targets with LLVM IR emission
Add an internal option -enable-early-lsr (default is false) to add loop
strength reduction pass to optimization pipeline for targets with LLVM IR
emission f.e. spir64. This pass normally runs at beginning of code gen, but
spir64 does not have code gen, so there is no way to run it now. This option
is intended to be used primarily for experiments.

Change-Id: I095e81aedfd3e0e836f062cf3ee466b1cf67a005
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/334536
Reviewed-by: Dmitriev, Serguei N <serguei.n.dmitriev@intel.com>
Tested-by: Dmitriev, Serguei N <serguei.n.dmitriev@intel.com>