[SYCL][Driver] Do not pass -fopenmp-simd to device compilations (#4098)
cmprllvm-26270: cherry-pick of 226edb8 and Intel customization

When -fopenmp-simd is used in conjunction with -fsycl, device
compilation should not be passed -fopenmp-simd.   Otherwise, it
leads to some common cases not being able to be compiled
(such as the definition of a function with a long double parameter).
This is already done for other OpenMP options.

In addition to the above, which was cherry-picked, a similar
change was also made for when -fiopenmp-simd is used.

Change-Id: I6c65ce037897ae663b542af01a74a3e1460f57ad
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/334041
Reviewed-by: Toguchi, Michael D <michael.d.toguchi@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>