[CMPLRLLVM-29795] Fix for assertion hit in DTransAnalysis on cpu2017ref/500
The DTransAnalysis is hitting an assertion due to a FreezeInst not being in
the list of accepted values for analyzing a value. This can be safely added
to the list of allowed instructions that produce the UnhandledSafety bit
because the cases that it is appearing in is not a candidate for DTrans.

Change-Id: Ifc9e914cf5ecdf761afab82c5051df1e27aad309
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/334092
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Zhang, Senran <senran.zhang@intel.com>
Tested-by: Zhang, Senran <senran.zhang@intel.com>