Prevent destructor replacement with base destructor when OpenMP offloading
The code that detemines which functions appear in the device output isn't
currently smart enough to jump over to the base destructor and will leave
it out causing a link failure.  This is already turned off for many cases.
If it becomes an issue we will need some possibly serious changes to the
code.

Change-Id: I71f251d3b3a74f2e3c6cff8e7b756b627eb8c0ac
Fixes: CMPLRLLVM-28607
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/334087
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Yu, Jennifer <jennifer.yu@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>