[SYCL] Omit TemplatedDecl from int-footer for a VarTemplateDecl (#4108)
The hook we had always added the VarDecl when it was visited, however
when evaluating a VarTemplateDecl, Clang creates a VarDecl to represent
the declaration as a child of the VarTemplateDecl. We don't need to
capture it in our collection since the VarTemplateSpecializationDecl
will provide the info we need.

If we DON'T do this, we end up having 2 specializations for the
specialization_id, one of which is invalid.

Change-Id: I5626fabf2f69a83672f0002f5e7894b2757c60af
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/334429
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Rice, Michael P <michael.p.rice@intel.com>
Reviewed-by: Aaron Ballman <aaron.ballman@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>