[DDTests] Strip out SExt/ZExt before performing isKnown*() checks in the presence of type mismatches.
These changes are required to fix CMPLRLLVM-29199 where we want to trigger
interchange for the hot loopnest.

The formed HIR after community pulldown has zext/sext inside the index expression
because of changes in ScalarEvolution's behavior.

It is better to make DDTests more robust in the presence of extensions than
trying to suppress them in ScalarEvolution.

Change-Id: I64804652e8b53c9126248f0bc1384c0dfde4cc6b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/334216
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>