[SYCL] Add sRGBA support (#4015)
Support for sRGBA linearization on read of RGBA images, exposing
this functionality on backends that have support for it. This is an
extension. The SYCL_EXT_ONEAPI_SRGB feature test macro and
aspect::ext_oneapi_srgb can be used to detect support in both
the runtime and the device. Documentation describing
limitations and usage is in llvm/sycl/docs/extensions/SRGB/

Signed-off-by: Chris Perkins <chris.perkins@intel.com>

This is cherri-pick for https://github.com/intel/llvm/pull/4015
Change-Id: I67a8fa493dbb3c5f35ddcc1c6c680168990bf014
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/334270
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>