[SYCL] Implement SYCL 2020 spec functionality: no propagation from function to the caller
In SYCL 1.2.1 spec, the attributes get propagated from device functions to a kernel.
The SYCL 2020 requirement mandating the avoidance of the propagation of all kernel attributes to the caller when used on a function.

Attributes that should not be propagated from device functions to a kernel to match with new SYCL 2020 spec.
1. scheduler_target_fmax_mhz
2. kernel_args_restrict
3. no_global_work_offset
4. max-work-group-size
5. max-global-work-dim
6. num-simd-work-items
7. reqd-sub-group-size
8. reqd-work-group-size
9. named_sub_group_size
10. sycl_explicit_simd

This patch
i. keeps the SYCL 1.2.1 spec functionality and propagates the attributes with the older SYCL mode(-sycl-std=2017)
iii. adds or updates tests to validate the propagating behavior with SYCL 1.2.1 and SYCL 2020 specs.

Fixes https://jira.devtools.intel.com/browse/CMPLRLLVM-16314
Cherry-pick from https://github.com/intel/llvm/pull/4084

Change-Id: Idc4d6f71d83e79e8caff05629b0a4433290b1bbe
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/334358
Reviewed-by: Manna, Soumi <soumi.manna@intel.com>
Tested-by: Manna, Soumi <soumi.manna@intel.com>