CMPLRLLVM-24359: Fix and document issues with stride updates for Fortran transpose.
This change set adds a StrideStores array for the TransposeCandidate class
and populates it either from the DopeVectorAnalyzer when a global variable
is transposed, or the DopeVectorInfo when a DopeVector based array is
transposed.  This is so the dope vector stride stores can be updated for
both in a uniform manner. It also documents the expected form of the
subscript calls and DopeVector dimensions after the transpose and fixes
the generation of the new strides for non-square arrays.

Change-Id: Ia8d1284d84fb18c412a80f68ce98b374ab7b1b4c
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/334914
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>