[DTrans - OpaquePtr] Fix DTrans base class to update metadata in cloned functions
The metadata attachments with DTrans type information should be updated to
reflect the type changes the same way regardless of whether the function is
remapped or is cloned, however the cloned functions were just copying the
metadata instead of performing the remapping of the information within them.
This updates the flag passed to the function cloning utility so that the
metadata is remapped instead of cloned. Adds tests to ensure both cloned
functions and non-cloned functions get updated.

Change-Id: I703f471aadf8076832bc8283f49a4b7c6e0a5760
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/334918
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>