[SYCL] Correct int-header emission with type aliases
Enums weren't correctly being printed when hidden by a type-alias, which
caused the integration header to not properly compile.  This was made
worse by the fact that the two enums in the test were exposed as type
aliases in the sycl headers for 2020.

This patch makes sure we always print the canonical name, rather than
having to try to figure out what type aliases need to be printed in
forward declarations.

Change-Id: Ie4691be0a150d9934377819d356c5616431d3650
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/335269
Reviewed-by: Bader, Alexey <alexey.bader@intel.com>
Reviewed-by: Andrews, Elizabeth <elizabeth.andrews@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>