[Driver][SYCL] Enable way to emit int-footer source to a specific dir (#4167)
During an -fsycl enabled build, we use an integration footer during the
host portion of the compilation.  This requires to create a temporary
file that is based on the original source file and contains the generated
footer.

Some build systems recognize this additional file that is being generated
and is not part of any expected generated dependency.  Introduce an option
that allows the user to specify the location in which this new source
file is created so the build system can properly track and recognize its
creation.  The new option is -fsycl-footer-path=<path>

Cherry-pick of https://github.com/intel/llvm/commit/155acd113a20045488cbeeff0510d90da86fd3d9
Addresses CMPLRLLVM-29930

Change-Id: I206dcbca8af99beddb26d6ab7801e65703f29bc3
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/335290
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Gu, Qichao <qichao.gu@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>