[Paropt] Emit opt-report remarks for redundant/simplified region clauses
This patch adds support for emitting opt-report messages for OpenMP work
regions in Paropt passes. Opt-report remarks now can be added to a work
region by OptReportBuilder in a similar way as it is done for loop

    ORBuilder(*W, WRegionList)
        .addRemark(OptReportVerbosity::Low, "remark text");

Opt-report metadata is attached to the region's directive entry call all
the way to Paropt transform pass which then moves it to the function
before removing work region's entry/exit calls from IR.

This functionality is now used by the shared-privatization pass for emitting
opt-report remarks for redundant or simplified work region clauses.

But it looks like opt-report metadata on functions in some cases gets lost
after Paropt transform pass, so some followup changes will be needed to
ensure that opt-report metadata is preserved.

Change-Id: I1544f4c9fe4fbbe0b143535eb398d023f664ad10
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/335089
Reviewed-by: Dmitriev, Serguei N <serguei.n.dmitriev@intel.com>
Tested-by: Dmitriev, Serguei N <serguei.n.dmitriev@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>