[DTrans - OpaquePtr] AOS-to-SOA pass transformation to update DTrans metadata on Alloca
When the AOS-to-SOA transformation converts a type, special handling is needed
for the DTrans metadata on Alloca instructions to handle the case where there
was a pointer-to-pointer type of the structure being transformed, because after
the transformation this type is made into a pointer to an integer index.
Fixes a couple of minor errors with changeset for bytegep handling.

Change-Id: I1ef9c8d2ad4d6cf41fc4ff44b479f9e0330f30bf
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/335309
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>