cmplrllvm-30108: fix assert on reference location is not set for opt-report.
the problem is for a variable with scalar non pointer type which
should in firstprivate clause, so the MapVarExprs is not getting update.

But the test case is the variable is in reduction clause, so it ends up in
map clause.

To fix this, adding the reference expression to MapVarExprs unconditionally.

Change-Id: I658338a7b65c1f8f23fb38115e3cee9469e79b71
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/335557
Tested-by: Yu, Jennifer <jennifer.yu@intel.com>
Reviewed-by: David Pagan <david.pagan@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>