CMPLRLLVM-24359: Check for out of bounds accesses for non-dope vector fields using -dtrans-outofboundsok=false.
LIT tests and testing are guarded for now under -dva-check-dtrans-outofboundsok
because I still see the issue on 527.cam4 where Brutus points to dope
vector constant propagation. I will resume debugging once this change set is
checked in (maybe even during the review).

The key idea is that with -dtrans-outofboundsok=false, a pointer to the
non-dopevector field is only an issue if once it is dereferenced it can be
used as a pointer and dereferenced again.

I have used a SetVector in the loads() interator for the DopeVectorFieldUse
so that the NestedDopeVectors always appear in the same order. This is very
helpful for triaging.

Change-Id: I9345a2af4ff7a39e405f3bf2845dbd55182419f5
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/335598
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>