CMPLRLLVM-30175: Fix -slib build breakage and associated LIT tests
My previous change broke the -slibs build. There is a dependency from the
Analysis component to the DTransAnalysis component, so if a lang rule based
option like DTransOutOfBoundOK needs to be used by both, it needs to be
defined in the Analysis component. I temporarily '#if 0'ed out the usage of
the DTransOutOfBoundsOK from Intel_DopeVectorAnalysis.cpp to fix the -slibs
build breakage, but two new LIT tests still failed.

I have added a new set of files Intel_LangRules.[h,cpp] to Analysis for
lang rules options. I will change the name to remove the "dtrans" in the
command line option when the driver changes are ready.

Change-Id: I3e3ffd7968694a0ba8d888bb602d1c66fcc1c41c
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/335824
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>