CMPLRLLVM-30118: Collect information from copy dope vectors
This change set introduces the data collection from dope
vectors that are copies from nested dope vectors. The
nested dope vectors aren't passed as arguments when
debug information is present. They are copied to a
local dope vector and then the local dope vector is
passed as an argument. These local dope vectores are
a new dope vector allocated, and then each dope
vector field is copied from the nested dope vector to
the new local dope vector.

The key idea here is to find all the places where the
nested dope vectors are copied, and then collect where
each fields are loaded. If the copy dope vector is
constant and pass validation, then we are going to merge
the load information collected into the original nested
dope vector. The transformation process will automatically
propagate the constants collected.

Change-Id: Id42326b49177527f931b5421f400841442988d86
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/335989
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>