CMPLRLLVM-24359: Fortran transpose should not be inhibited by string operation
The previous change set had a bug in it which was caught by Prasad, but also
revealed a needed piece of analysis. A non-dope vector field which was a
character string had its address passed into a Fortran concatenation table
and was then passed to for_concat(). The escape analysis must be extended
to handle this case.

Change-Id: I8c7e5bc6fefd790c61ea5e2a1ea9360e43843074
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/336108
Tested-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>