CMPLRLLVM-30118: Collect information from copy dope vectors
This change set introduces the data collection from dope
vectors that are copies from nested dope vectors. The
nested dope vectors aren't passed as arguments when
debug information is present. They are copied to a
local dope vector and then the local dope vector is
passed as an argument. These local dope vectores are
a new dope vector allocated, and then each dope
vector field is copied from the nested dope vector to
the new local dope vector.

The key idea here is to find all the places where the
nested dope vectors are copied, and then collect where
each fields are loaded. If the copy dope vector is
constant and pass validation, then we are going to merge
the load information collected into the original nested
dope vector. The transformation process will automatically
propagate the constants collected.

These are the same changes as
https://git-amr-2.devtools.intel.com/gerrit/#/c/335989/
but for xmain-rel. They also contain a minor fix that
changes storing the load instructions from SmallPtrSet
to SetVector. That is to produce deterministic results
when printing the debug data. The fix is part of this
change set:
https://git-amr-2.devtools.intel.com/gerrit/#/c/335598/

Change-Id: I053847ad951c1f132320b12238aaf7de63e29765
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/336245
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>