[SYCL] Add a workaround for a Level Zero batching issue (#4268)
Consider the following scenario on Level Zero:
1. Kernel A, which uses buffer A, is submitted to queue A.
2. Kernel B, which uses buffer B, is submitted to queue B.
3. queueA.wait().
4. queueB.wait().

DPCPP runtime used to treat unmap/write commands for buffer A/B as host
dependencies (i.e. they were waited for prior to enqueueing any command
that's dependent on them). This allowed Level Zero plugin to detect that
each queue is idle on steps 1/2 and submit the command list right away.
This is no longer the case since we started passing these dependencies in
an event waitlist and Level Zero plugin attempts to batch these commands,
so the execution of kernel B starts only on step 4. This workaround
restores the old behavior in this case until this is resolved.

Change-Id: I8c215062ff0b1407b7511e2b9e36e08db8ff5ea9
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/336272
Reviewed-by: Semenov, Sergey <sergey.semenov@intel.com>
Tested-by: Semenov, Sergey <sergey.semenov@intel.com>
Reviewed-by: Kurkov, Vasiliy A <vasiliy.a.kurkov@intel.com>