[CMPLRLLVM-29133] Generalize CTLZ detection in LoopIdiomRecognize
This change builds on 8413fab295492a5fc19bdd5a9193ecd864a6d1b0 to even
more significantly expand the range of loops recognized as following the
CTLZ (and to a lesser extent CTTZ) idiom. Besides the loops already
handled, the pass now supports:

* Loop conditions of the form "< (1<<N)" (or equivalent comparisons)
  either before or after the shift
* The same types of comparisons as zero-trip guards
* Replacing uses of either or both pre- and post-increment loop count
  values in all recognized cases

This ought to be good enough to keep us out ahead of whatever instcombine
decides to do with these CTLZ loops for a while.

Change-Id: I9552bbc65a89e96a362dba78b7255b9800ef6a0d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/336197
Reviewed-by: Lin Jr, Chang-Sun <chang-sun.lin.jr@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Mikhail Lychkov <mikhail.lychkov@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/337276
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Kurkov, Vasiliy A <vasiliy.a.kurkov@intel.com>