[VPlan] Fix assert when -vplan-enable-peeling is on by default
Assertion `StaticPeel->peelCount() && "unexpected zero peel count"' failed
is reached when peeling is turned on. For the current logic, it seems that
getPreferredPeeling(VF) = nullptr should indicate no peeling, but
VPlanStaticPeeling{1} is used for setPreferredPeeling. I matched the same
for the case when there is a static peeling variant and peelcount is found
to be 0. However, it seems that we should be using VPlanStaticPeeling{0} for
both cases if I understand correctly. Please comment.

Change-Id: I5a42e8feb1d3760395e5eb77e73ab25b3b803f7d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/336356
Reviewed-by: Masten, Matt <matt.masten@intel.com>
Tested-by: Masten, Matt <matt.masten@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>