[VPlan] Fix assert when peeling is enabled
Alignment analysis gathers memrefs for the purposes of determining
which of them are congruent. HIR interfaces used in this process
require these memrefs to be of the same type, otherwise nullptr
is returned. As such, we hit the assert, Assertion `Diff &&
"CanonExprs are not mergeable"' failed because the two memrefs
being considered are not of the same type and CanonExprUtils::
cloneAndSubtract returns nullptr. This patch lets the callers
handle the results accordingly and removes the assert within the
getMinusExprImpl function itself. In this case, if two memrefs
are not of the same type, then it is assumed they are not congruent.

Change-Id: I03090dcdfda0e3366aaf2cfa111260e50f9e5ec2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/336475
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>