[VPlan] Fix assert when enabling peeling by default
Hit assert Assertion `NewVal && "Value::replaceUsesWithIf(<null>)
is invalid!"' failed because live-in for scalar peel loop does not
have an original incoming value. Thus, there will not be any uses
of that value to replace. This patches fixes this by checking to
make sure that live-ins have such values before trying to replace
them. Additional test modifications were made to account for the
fact that peeling has to be turned on to hit the assert and the
CFG has changed.

Change-Id: I75cfc32c26ede098f8d79a0f3e55b68d1d77fd06
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/336487
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>