CMPLRLLVM-30322: Fix for multiversioning after pulldown
This change set fixes the function multiversioning after pulldown.
If an instruction is used in a freeze instruction, then all the
users for this instruction will be replaced with the freeze
instruction by InstCombine. The instruction replacement was
added by the community in D106233. We need to collect the freeze
instruction and keep tracing the users for it too.

This change set restores 3.71% for CPU2017/500.perlbench.

Change-Id: I50ebc2832462c112bd38b3f057ccce4d6622f0cd
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/336591
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>