CMPLRLLVM-30432: Remove code unfriendly to opaque pointers in Intel_TileMVInlMarker.cpp
Extend and use inferPtrElementType() to test for tile argument pointer to double.
inferPtrElementType() is now in its own file so it can be invoked in mutiple
places.

Change IsCanonicalFortranRead() to work the same way with and without
bitcasts. To simplify the code, I added an additional test to check for a
single use for the bitcast.

I verified the tests in Transforms/Inline and 549.fotonik still work after this
change. I tried to convert the LIT tests to use opaque pointers, but the verifier
dies when checking an opaque pointer version of SubscriptInst. Once this is fixed,
I will go back and add the opaque pointer LIT tests for this.

Change-Id: Ifd8da886f21efa5080f82ef5737447eb61638598
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/336615
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>