[SYCL] Correctly form the integration header for non-base-ascii char (#4287)
Unnamed kernels use the mangled name, which in some cases can contain an
arbitrary 'char' byte when unpacked using the integer_seq in the header.

Because identifiers can be a pretty massive set of unicode characters,
make sure that we can encode them properly.  We do this by being really
conservative and limiting ourselves to only printing as a
single-quote-delimited character literal when it is one of the
base-ascii printable characters, otherwise emit it as a signed integer
literal instead.

Change-Id: I288a7a88f6802d8bd66d91a829358230a1f840b1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/336721
Reviewed-by: Andrews, Elizabeth <elizabeth.andrews@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/336862
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Rice, Michael P <michael.p.rice@intel.com>
Reviewed-by: Mikhail Lychkov <mikhail.lychkov@intel.com>