[SYCL] Delay kernel naming until after all pending instantiations (#4296)
If we do Kernel naming too early, the kernel names haven't
"settled" down enough for our names to be correct. This can result in a
compile failure where two kernels end up getting the same name without a
good reason.  This patch separates the naming of a kernel from the
constructing of a kernel, and calls it 'later'.

Change-Id: I9dbc63350dcb7f4baf4ebe9e13be0e3f8f5b9bb1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/336724
Reviewed-by: Keane, Erich <erich.keane@intel.com>
Tested-by: Keane, Erich <erich.keane@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/336891