[loopopt] OptReport remarks update for HIROptVarPredicate
Remark updated for the generic case. Remark added for the peeling case,
and the case when the loop is optimized away (in those cases the generic
remark is not printed). Comments based on addOrigin are now only printed
if loops are multiversioned, i.e., there are more than one output loops.
Lit-tests updated accordingly.

Change-Id: I6079555a15460438db963bbfe632525412ef129f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/336760
Reviewed-by: Fawad Ismail <fawad.ismail@intel.com>
Tested-by: Fawad Ismail <fawad.ismail@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>