CMPLRLLVM-30417: Fix regression in 647.cam4 due to putting code under -dtrans-outofboundsok=false.
We got a 2% regression in 627.cam4 when I put some code for analyzing
non-dope-vector fields in a structure with dope vectors under
-dtrans-outofboundsok=false. It turns out that 3 other global variables
in 627.cam4 were using this code to their advantage to get constant dope
vector fields. So, I have restored the code needed to do this, while
leaving the rest of it still under -dtrans-outofboundsok=false.

Change-Id: Icfef10a04777f8f370889941a580c0763b592d4d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/336787
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>