[VPlan][Peeling] Add separate command line switches to enable peeling for LLVM IR and HIR paths
We are currently using one command line switch to enable alignment analysis
as well as peeling profitability analysis for LLVM IR and HIR paths. Alignment
analysis and peeling for HIR path is currently WIP and trying to make the
command line switch that controls alignment analysis and peeling on by default
exposes issues in the HIR path. To avoid the HIR specific issues, separate
switches are added for the LLVM IR and HIR paths so that the effort to enable
peeling by default in the LLVM IR path is not blocked by issues in the HIR
path.

Jira :: CMPLRLLVM-30026

Change-Id: Ifeebae89aa269c7ac5e3edbbb8c4ae130326527a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/336811
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Masten, Matt <matt.masten@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>