[DTrans - OpaquePtr] PtrTypeAnalyzer handling of metadata inserted by DTrans transformations
Some DTrans transformation mark special metadata annotations on instructions
that would not normally have DTrans type metadata and not obviously analyzable
from the IR. For example, the AOS-to-SOA transformation allocates a large block
of memory for the structure of arrays data structure, which is then subdivided
into individual arrays. The instructions that get the address of the
subdivisions are annotated to allow subsequent DTrans passes to determine the
safety of the memory accesses.

Change-Id: I953d548fe67dfd8f1db31c8de40ed777f2498acc
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/336904
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>