Disable "Transform sin and cos calls" for SYCL pipeline
Built-ins, used by the transformation are not documented by any SYCL or
OpenCL specification and therefore they can't be reliable used.

The alternative approach would be to provide definition of those
built-ins through "devicelib" component, but we likely won't be able to
put a high-performance implementation there, so as a short/medium term
approach it seems more reasonable to disable the transformation,
improving the compilation time a bit.

This fixes CMPLRLLVM-30382

Change-Id: I21cf0b6f0846d6d7abcf3f83c24fa307b5cd9aaf
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/336965
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Sachkov, Alexey <alexey.sachkov@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>