CMPLRLLVM-30521: Remove code unfriendly to opaque pointers in IP cloning files.
This is a first change set that removes references to getPointerElementType()
and PointerType::getElementType(). There are too many changes here to attempt
dealing with BitCastInsts in the same change set.

I had to change some of the LIT tests, because they had Arguments without
uses, and with opaque pointers we will need to infer the pointer element
type of an Argument from its uses. The original benchmarks from which those
LIT tests were derived did have Arguments with uses.

Some of the function argument and return types in the files have been
"tightened up" (I have changed the types to derived types) because not all
types have easily inferred or computed pointer element types when we have
opaque pointers. This makes it easier to see what types are actually passed
to certain functions so that we can more easily derive pointer element
types when needed.

In one case, I used the corresponding formal rather than actual type of a
CallBase to get the pointer element type.

Change-Id: If3fe91367300356d5c055d49a42408b6d7aa58df
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/337015
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>